requirements.txt
Django==4.1
djangorestframework==3.14.0
pydantic==2.5.3
ebaysdk==2.2.0
django-cors-headers==4.3.1
whitenoise==6.6.0
